Refactor legacy task routes
